requirements.txt
flask>=2.0
flask-compress>=1.13
requests>=2.25
psutil>=5.8
nbtlib>=2.0
apscheduler>=3.9
pytz>=2023.3
SQLAlchemy>=1.4
alembic>=1.11
argon2-cffi>=21.3.0
pyotp>=2.8.0
paramiko>=3.3.1
boto3
PyYAML
Pillow
flask-limiter
prometheus-client
